+++ plugins/mysql-adapter/src/main/java/org/polypheny/db/adapter/jdbc/MysqlSourcePlugin.java
@@ -20,7 +20,6 @@ package org.polypheny.db.adapter.jdbc;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.polypheny.db.adapter.AdapterManager;
 import org.polypheny.db.adapter.DeployMode;
@@ -30,13 +29,8 @@ import org.polypheny.db.adapter.annotations.AdapterSettingInteger;
 import org.polypheny.db.adapter.annotations.AdapterSettingList;
 import org.polypheny.db.adapter.annotations.AdapterSettingString;
 import org.polypheny.db.adapter.jdbc.sources.AbstractJdbcSource;
-import org.polypheny.db.catalog.entity.allocation.AllocationTableWrapper;
-import org.polypheny.db.catalog.entity.logical.LogicalTableWrapper;
-import org.polypheny.db.catalog.entity.physical.PhysicalEntity;
-import org.polypheny.db.catalog.entity.physical.PhysicalTable;
 import org.polypheny.db.plugins.PluginContext;
 import org.polypheny.db.plugins.PolyPlugin;
-import org.polypheny.db.prepare.Context;
 import org.polypheny.db.sql.language.dialect.MysqlSqlDialect;
 
 @SuppressWarnings("unused")
@@ -94,24 +88,6 @@ public class MysqlSourcePlugin extends PolyPlugin {
 
         public MysqlSource( final long storeId, final String uniqueName, final Map<String, String> settings, final DeployMode mode ) {
             super( storeId, uniqueName, settings, mode, "org.mariadb.jdbc.Driver", MysqlSqlDialect.DEFAULT, false );
-        }
-
-
-        @Override
-        public List<PhysicalEntity> createTable( Context context, LogicalTableWrapper logical, AllocationTableWrapper allocation ) {
-            PhysicalTable table = adapterCatalog.createTable(
-                    logical.table.getNamespaceName(),
-                    logical.table.name,
-                    logical.columns.stream().collect( Collectors.toMap( c -> c.id, c -> c.name ) ),
-                    logical.table,
-                    logical.columns.stream().collect( Collectors.toMap( t -> t.id, t -> t ) ),
-                    logical.pkIds, allocation );
-
-            JdbcTable physical = currentJdbcSchema.createJdbcTable( table );
-
-            adapterCatalog.replacePhysical( physical );
-
-            return List.of( physical );
         }
